requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
anthropic==0.7.8
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
jinja2==3.1.2
numpy==1.24.3
sentence-transformers==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.0
httpx==0.25.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-magic==0.4.27 